Move core state into module
