+++ src/immutable/state/index.ts
@@ -2,16 +2,15 @@ import { Record } from 'immutable';
 import { RouterState } from 'connected-react-router';
 
 import { AdminStateRecord } from './admin';
-import { CoreStateRecord } from './core';
 import { UsersStateRecord } from './users';
 import { ADMIN_NAMESPACE } from '../../modules/admin/constants';
 import { CORE_NAMESPACE } from '../../modules/core/constants';
+import { CoreStateRecord } from '../../modules/core/state/index';
 import { DASHBOARD_NAMESPACE } from '../../modules/dashboard/constants';
 import { DashboardStateRecord } from '../../modules/dashboard/state/index';
 import { USERS_NAMESPACE } from '../../modules/users/constants';
 
 export * from './admin';
-export * from './core';
 export * from './users';
 
 export interface RootStateProps {
