+++ src/modules/core/state/index.ts
@@ -1,15 +1,19 @@
 import { Record, Map as ImmutableMap } from 'immutable';
 
-import { Connection, ConnectionRecord } from '../../Connection';
-import { FetchableData, FetchableDataRecord } from '../../FetchableData';
+import {
+  Connection,
+  ConnectionRecord,
+  FetchableData,
+  FetchableDataRecord,
+  GasPrices,
+  GasPricesRecord,
+  Network,
+  NetworkRecord,
+} from '~immutable/index';
 import { CoreTransactions, CoreTransactionsRecord } from './CoreTransactions';
-import { GasPrices, GasPricesRecord } from './GasPrices';
-import { Network, NetworkRecord } from './Network';
 import { CoreMessages, CoreMessagesRecord } from './Messages';
 
-export * from './GasPrices';
 export * from './CoreTransactions';
-export * from './Network';
 export * from './Messages';
 
 export type IpfsDataType = ImmutableMap<string, FetchableDataRecord<string>>;
@@ -23,6 +27,7 @@ export type CoreStateProps = {
   transactions: CoreTransactionsRecord;
 };
 
+// FIXME use constants for state everywhere
 export class CoreStateRecord extends Record<CoreStateProps>({
   connection: Connection(),
   gasPrices: GasPrices(),
